Remove job arguments from perform log
In general, I think this actually needs more improvement. The logging on
the `perform` method seems to be very verbose and not super beneficial
in practice.

The actual reason why I culled the job arguments are simple.
Implementors of jobs may pass items like passwords, session tokens, and
other things to quebert for job processing as arguments to jobs.
Unfortuantely, this log entry guarantees that all of those arguments are
pushed to whatever logging system is in service. In many cases this is
internal, but in some caes, a user with a third party like Datadog may
accidentally send sensitive information to that third party.

I'm open to alternatives on this. In particular, I considered, but did
not implement something like a `quiet_perform` or changing the arguments
to perform to take a quiet argument as both seem like they would
introduce API breaking changes. Cutting off the debug output here would
make sense as any job that actually needs to send debug output can
easily do so when it executes. It leaves the decision up to the person
implementing the job, not implementing the job processor.